+++ src/app/layout/upload/upload.module.ts
@@ -2,9 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 // import { ReactiveFormsModule } from '@angular/forms';
 import {FormsModule} from '@angular/forms';
+import {  NgbTooltipModule, NgbTabsetModule } from '@ng-bootstrap/ng-bootstrap';
+
 
 import { PageHeaderModule } from '../../shared';
-
 import { UploadRoutingModule } from './upload-routing.module';
 import { UploadComponent } from './upload.component';
 
@@ -14,6 +15,8 @@ import { UploadComponent } from './upload.component';
     PageHeaderModule,
     // ReactiveFormsModule,
     FormsModule,
+    NgbTooltipModule,
+    NgbTabsetModule,
     UploadRoutingModule,
   ],
   declarations: [UploadComponent]